renamed utils.ts to index.ts and updated imports accordingly
